[arch][x86-64] Bug fix on bit operation and others
1. Add a memory barrier to context switch on smp system.
2. Fix 64-bit bit operation failure
    Sizeof long takes 4 bytes in 32-bit and 8 bytes in 64-bit
    Use BITMAP_BIT_IN_WORD to replace BITMAP_BIT_IN_WORD can
    deal with both 32-bit and 64-bit
3. Remove the build of new.cpp in lib/libc-trusty which is
    duplicated with lib/libstdc++-trusty
4. Add x86/toolchain.mk which is depended by x86-64

+++ arch/x86-64/include/arch/arch_ops.h
@@ -95,9 +95,8 @@ static inline int atomic_swap(volatile int *ptr, int val)
 	return val;
 }
 
-
-static inline int atomic_and(volatile int *ptr, int val) { return _atomic_and(ptr, val); }
-static inline int atomic_or(volatile int *ptr, int val) { return _atomic_or(ptr, val); }
+static inline int atomic_and(volatile int *ptr, long val) { return _atomic_and(ptr, val); }
+static inline int atomic_or(volatile int *ptr, long val) { return _atomic_or(ptr, val); }
 
 static inline int atomic_cmpxchg(volatile int *ptr, int oldval, uint64_t newval)
 {
